+++ core/constants.js
@@ -11,6 +11,10 @@ const PERMISSION_WRITE = 'write'
 const PERMISSION_MANAGE = 'manage'
 const ORDERED_PERMISSIONS = [PERMISSION_READ, PERMISSION_WRITE, PERMISSION_MANAGE]
 
+const SPECIAL_POST_TYPE_ANNOUNCEMENT = 'announcement'
+const SPECIAL_POST_TYPE_ARTICLE = 'article'
+const SPECIAL_POST_TYPES = [SPECIAL_POST_TYPE_ANNOUNCEMENT, SPECIAL_POST_TYPE_ARTICLE]
+
 module.exports = {
   // Settings
   SETTING_DB_VERSION: 'db_version',
@@ -25,7 +29,10 @@ module.exports = {
   ORDERED_PERMISSIONS,
 
   // Posts
-  SPECIAL_POST_TYPE_ANNOUNCEMENT: 'announcement',
+  SPECIAL_POST_TYPE_ANNOUNCEMENT,
+  SPECIAL_POST_TYPE_ARTICLE,
+  SPECIAL_POST_TYPES,
+  REQUIRED_ARTICLES: ['feedback'],
   ALLOWED_POST_TAGS: [ 'b', 'i', 'em', 'strong', 'a', 'u', 'br' ],
   ALLOWED_POST_ATTRIBUTES: { 'a': [ 'href' ] },